+++ src/backend_aarch64.hpp
@@ -131,10 +131,10 @@ class Aarch64Backend : public Backend
 public:
     Aarch64Backend();
     virtual size_t reusingPreferences(const Syntop& a_op, const std::set<size_t>& undefinedArgNums) const override final;
-    virtual size_t spillSpaceNeeded(const Syntop& a_op, int basketNum) const override final;
+    virtual int spillSpaceNeeded(const Syntop& a_op, int basketNum) const override final;
     virtual std::set<size_t> getUsedRegistersIdxs(const Syntop& a_op, int basketNum, uint64_t flagmask = BinTranslation::Token::T_INPUT | BinTranslation::Token::T_OUTPUT) const override final;
-    virtual void getStackParameterLayout(const Syntfunc& a_func, const std::vector<size_t> (&regParsOverride)[RB_AMOUNT], std::map<RegIdx, size_t> (&parLayout)[RB_AMOUNT]) const override final;
-    virtual size_t stackGrowthAlignment(size_t stackGrowth) const override final;
+    virtual void getStackParameterLayout(const Syntfunc& a_func, const std::vector<int> (&regParsOverride)[RB_AMOUNT], std::map<RegIdx, int> (&parLayout)[RB_AMOUNT]) const override final;
+    virtual int stackGrowthAlignment(int stackGrowth) const override final;
     virtual void writeCallerPrologue(Syntfunc& prog, int stackGrowth) const override final;
     virtual void writeCallerEpilogue(Syntfunc& prog, int stackGrowth) const override final;
     virtual Arg getSParg() const override final;
